Outsourced in utils.py
